LayoutTests/http/tests/security/set-domain-remove-subdomain-for-ip-address-expected.txt

Test whether a page loaded straight from an IP address can set document.domain and wrongly treat IP address octets as subdomains.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.domain = '0.0.1' threw exception SecurityError (DOM Exception 18): The operation is insecure..
PASS document.domain is "127.0.0.1"
PASS successfullyParsed is true

TEST COMPLETE
